requirements.txt
psycopg2
Django~=4.1.5
djangorestframework
gunicorn==20.1.0
sqlparse==0.4.2
psycopg2-binary==2.9.3